Restructure  tool agent tests
